+++ packages/imageGenerationService/src/index.js
@@ -1,20 +1,3 @@
-import { serverGateways } from 'vizhub-server-gateways';
-import { UpdateImages } from 'datavis-tech-use-cases';
-import { generateImages } from './generateImages';
+import { startService } from './service';
 
-const gateways = Object.assign({}, serverGateways(), {
-  imageGeneratorGateway: {
-    generateImages
-  }
-});
-
-const updateImages = new UpdateImages(gateways);
-
-const updateImagesInterval = 12 * 1000;
-
-console.log('Image generation service starting...');
-
-setInterval(() => {
-  //console.log('Checking for visualizations in need of images..');
-  updateImages.execute();
-}, updateImagesInterval);
+startService();
